requirements.txt
python-dotenv
sqlanydb
